+++ CHANGELOG.rst
@@ -19,6 +19,9 @@ Added
 
 Changed
 -------
+- improved the online learning experience on the CLI
+- made finetuning during online learning optional (use ``--finetune`` if
+  you want to enable it)
 
 Removed
 -------